+++ feature/namecard/src/main/java/com/wepli/feature/namecard/main/navigation/NameCardNavigation.kt
@@ -1,9 +1,9 @@
-package com.wepli.feature.namecard.navigation
+package com.wepli.feature.namecard.main.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.wepli.feature.namecard.NameCardScreenRoute
+import com.wepli.feature.namecard.main.NameCardScreenRoute
 import com.wepli.navigator.feature.namecard.NameCardRoute
 import extensions.enterAnimation
